+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - PR #574 dev env update
 - PR #580 Changed hardcoded test output file to a generated tempfile file name
 - PR #595 Updates to use the new RMM Python reinitialize() API
+- PR #625 use destination instead of target when adding edgelist
 
 # cuGraph 0.10.0 (16 Oct 2019)